Reorganize files in new folders
Improve folder structure
