+++ Tests/testLyrics.py
