+++ bot.py
@@ -3,7 +3,7 @@ import sys
 import ssl
 import time
 import config
-from chess_game import ChessManager
+from Classes.chess_game import ChessManager
 from pymongo.mongo_client import MongoClient
 from Commands import (bot_info, date, groq_command, help_ascii, ping, help_chess, source_code, play_chess, ro, r960, help_ro, pyramid, slow_pyramid,
                       news, help_news, daily, roulette, balance, leaderboard, help, shop, timeout, trophies, gemini, gemini2,
